pyproject.toml
[tool.poetry]
name = "quizzing"
version = "0.1.0"
description = ""
authors = ["Daniel Fonnegra <dfonnegrag@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
bcrypt = "^4.1.3"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
